refactor: move private stuff to -private folder
