+++ addon/-private/config.js
